fakultas policy

+++ app/Http/Controllers/FakultasController.php
@@ -77,6 +77,10 @@ class FakultasController extends Controller
     public function update(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
+        // cek apakah user memiliki izin untuk mengupdate fakultas
+        if($request->user()->cannot('update', $fakultas)) {
+            abort(403, 'Unauthorized action.');
+        }
         // validasi input
         $input = $request->validate([
             'nama' => 'required',
@@ -93,11 +97,14 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
         // dd($fakultas);
-
+        // cek apakah user memiliki izin untuk menghapus fakultas
+        if($request->user()->cannot('delete', $fakultas)) {
+            abort(403, 'Unauthorized action.');
+        }
         // Hapus data fakultas
         $fakultas->delete();
         // Redirect ke route fakultas.index
